requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
